+++ src/main.rs
@@ -2,19 +2,36 @@ use log::{debug, error};
 use std::io::Write;
 use structopt::StructOpt;
 
-#[derive(structopt::StructOpt, Debug)]
+#[derive(structopt::StructOpt, Debug, PartialEq)]
 struct CommonOpts {
-    /// Restart the wrapped program if it exits with 0
-    #[structopt(long)]
-    restart_ok: bool,
+    /// Exit codes that should be considered successful (comma-separated)
+    #[structopt(long, value_name = "codes", default_value = "0", use_delimiter(true))]
+    pass: Vec<i32>,
 
-    /// Do not restart the wrapped program if it exits with a nonzero code
+    /// Always restart the command regardless of the exit code
+    #[structopt(
+        long,
+        conflicts_with("no-restart"),
+        conflicts_with("restart-if"),
+        conflicts_with("restart-if-not")
+    )]
+    restart: bool,
+
+    /// Never restart the command regardless of the exit code
     #[structopt(long)]
-    no_restart_err: bool,
+    no_restart: bool,
+
+    /// Restart the command if the exit code is one of these (comma-separated)
+    #[structopt(long, value_name = "codes", use_delimiter(true))]
+    restart_if: Vec<i32>,
+
+    /// Restart the command if the exit code is not one of these (comma-separated)
+    #[structopt(long, value_name = "codes", use_delimiter(true))]
+    restart_if_not: Vec<i32>,
 
     /// How long to wait in milliseconds between sending the wrapped process
     /// a ctrl-C event and forcibly killing it
-    #[structopt(long, default_value = "3000")]
+    #[structopt(long, default_value = "3000", value_name = "ms")]
     stop_timeout: u64,
 
     /// Command to run as a service
@@ -22,7 +39,7 @@ struct CommonOpts {
     command: Vec<String>,
 }
 
-#[derive(structopt::StructOpt, Debug)]
+#[derive(structopt::StructOpt, Debug, PartialEq)]
 enum Subcommand {
     #[structopt(about = "Add a new service")]
     Add {
@@ -33,7 +50,9 @@ enum Subcommand {
         #[structopt(long)]
         name: String,
     },
-    #[structopt(about = "Run a command as a service; only works when launched by the Windows service manager")]
+    #[structopt(
+        about = "Run a command as a service; only works when launched by the Windows service manager"
+    )]
     Run {
         #[structopt(flatten)]
         common: CommonOpts,
@@ -44,7 +63,7 @@ enum Subcommand {
     },
 }
 
-#[derive(structopt::StructOpt, Debug)]
+#[derive(structopt::StructOpt, Debug, PartialEq)]
 #[structopt(
     name = "shawl",
     about = "Wrap arbitrary commands as Windows services",
@@ -60,18 +79,16 @@ fn prepare_logging(console: bool) -> Result<(), Box<std::error::Error>> {
     log_file.pop();
     log_file.push("shawl.log");
 
-    let mut loggers: Vec<Box<simplelog::SharedLogger>> = vec![
-        simplelog::WriteLogger::new(
-            simplelog::LevelFilter::Debug,
-            simplelog::ConfigBuilder::new()
-                .set_time_format_str("%Y-%m-%d %H:%M:%S")
-                .build(),
-            std::fs::OpenOptions::new()
-                .create(true)
-                .append(true)
-                .open(log_file)?,
-        )
-    ];
+    let mut loggers: Vec<Box<simplelog::SharedLogger>> = vec![simplelog::WriteLogger::new(
+        simplelog::LevelFilter::Debug,
+        simplelog::ConfigBuilder::new()
+            .set_time_format_str("%Y-%m-%d %H:%M:%S")
+            .build(),
+        std::fs::OpenOptions::new()
+            .create(true)
+            .append(true)
+            .open(log_file)?,
+    )];
 
     if console {
         loggers.push(
@@ -81,7 +98,8 @@ fn prepare_logging(console: bool) -> Result<(), Box<std::error::Error>> {
                     .set_time_format_str("")
                     .build(),
                 simplelog::TerminalMode::default(),
-            ).expect("Unable to create terminal logger")
+            )
+            .expect("Unable to create terminal logger"),
         );
     }
 
@@ -99,11 +117,31 @@ fn add_service(name: String, opts: CommonOpts) -> Result<(), ()> {
         "--stop-timeout".to_string(),
         opts.stop_timeout.to_string(),
     ];
-    if opts.restart_ok {
-        shawl_args.push("--restart-ok".to_string());
+    if opts.restart {
+        shawl_args.push("--restart".to_string());
     }
-    if opts.no_restart_err {
-        shawl_args.push("--no-restart-err".to_string());
+    if opts.no_restart {
+        shawl_args.push("--no-restart".to_string());
+    }
+    if !opts.restart_if.is_empty() {
+        shawl_args.push("--restart-if".to_string());
+        shawl_args.push(
+            opts.restart_if
+                .iter()
+                .map(|x| x.to_string())
+                .collect::<Vec<String>>()
+                .join(","),
+        );
+    }
+    if !opts.restart_if_not.is_empty() {
+        shawl_args.push("--restart-if-not".to_string());
+        shawl_args.push(
+            opts.restart_if_not
+                .iter()
+                .map(|x| x.to_string())
+                .collect::<Vec<String>>()
+                .join(","),
+        );
     }
 
     let output = std::process::Command::new("sc")
@@ -135,6 +173,26 @@ fn add_service(name: String, opts: CommonOpts) -> Result<(), ()> {
     }
 }
 
+fn should_restart_exited_command(
+    code: &i32,
+    restart: bool,
+    no_restart: bool,
+    restart_if: &Vec<i32>,
+    restart_if_not: &Vec<i32>,
+) -> bool {
+    if !restart_if.is_empty() {
+        restart_if.contains(code)
+    } else if !restart_if_not.is_empty() {
+        !restart_if_not.contains(code)
+    } else {
+        restart || !no_restart && *code != 0
+    }
+}
+
+fn should_restart_terminated_command(restart: bool, _no_restart: bool) -> bool {
+    restart
+}
+
 #[cfg(windows)]
 fn main() -> Result<(), Box<std::error::Error>> {
     let cli = Cli::from_args();
@@ -161,7 +219,7 @@ fn main() -> Result<(), Box<std::error::Error>> {
                 // tried to run it directly, so try showing them the error:
                 println!("Failed to run the service:\n{:#?}", e);
                 std::process::exit(1)
-            },
+            }
         },
     }
     debug!("Finished successfully");
@@ -175,8 +233,7 @@ fn main() {
 
 enum ProcessStatus {
     Running,
-    Success(i32),
-    Failure(i32),
+    Exited(i32),
     Terminated,
 }
 
@@ -184,8 +241,7 @@ fn check_process(child: &mut std::process::Child) -> Result<ProcessStatus, Box<s
     match child.try_wait() {
         Ok(None) => Ok(ProcessStatus::Running),
         Ok(Some(status)) => match status.code() {
-            Some(0) => Ok(ProcessStatus::Success(0)),
-            Some(code) => Ok(ProcessStatus::Failure(code)),
+            Some(code) => Ok(ProcessStatus::Exited(code)),
             None => Ok(ProcessStatus::Terminated),
         },
         Err(e) => Err(Box::new(e)),
@@ -323,16 +379,15 @@ mod service {
                                         break;
                                     }
                                 }
-                                Ok(crate::ProcessStatus::Success(code))
-                                | Ok(crate::ProcessStatus::Failure(code)) => {
+                                Ok(crate::ProcessStatus::Exited(code)) => {
                                     info!(
                                         "Command exited after {:?} ms with code {:?}",
                                         start_time.elapsed().as_millis(),
                                         code
                                     );
-                                    service_exit_code = match code {
-                                        0 => ServiceExitCode::NO_ERROR,
-                                        x => ServiceExitCode::ServiceSpecific(x as u32),
+                                    service_exit_code = match opts.pass.contains(&code) {
+                                        true => ServiceExitCode::NO_ERROR,
+                                        false => ServiceExitCode::ServiceSpecific(code as u32),
                                     };
                                     break;
                                 }
@@ -351,18 +406,19 @@ mod service {
 
                 match crate::check_process(&mut child) {
                     Ok(crate::ProcessStatus::Running) => (),
-                    Ok(crate::ProcessStatus::Success(_)) => {
-                        info!("Command finished successfully");
-                        service_exit_code = ServiceExitCode::NO_ERROR;
-                        match opts.restart_ok {
-                            true => break 'inner,
-                            false => break 'outer,
-                        }
-                    }
-                    Ok(crate::ProcessStatus::Failure(code)) => {
-                        info!("Command failed with code {}", code);
-                        service_exit_code = ServiceExitCode::ServiceSpecific(code as u32);
-                        match !opts.no_restart_err {
+                    Ok(crate::ProcessStatus::Exited(code)) => {
+                        info!("Command exited with code {:?}", code);
+                        service_exit_code = match opts.pass.contains(&code) {
+                            true => ServiceExitCode::NO_ERROR,
+                            false => ServiceExitCode::ServiceSpecific(code as u32),
+                        };
+                        match crate::should_restart_exited_command(
+                            &code,
+                            opts.restart,
+                            opts.no_restart,
+                            &opts.restart_if,
+                            &opts.restart_if_not,
+                        ) {
                             true => break 'inner,
                             false => break 'outer,
                         }
@@ -371,7 +427,10 @@ mod service {
                         info!("Command was terminated by a signal");
                         service_exit_code =
                             ServiceExitCode::Win32(winapi::shared::winerror::ERROR_PROCESS_ABORTED);
-                        match !opts.no_restart_err {
+                        match crate::should_restart_terminated_command(
+                            opts.restart,
+                            opts.no_restart,
+                        ) {
                             true => break 'inner,
                             false => break 'outer,
                         }
@@ -397,5 +456,385 @@ mod service {
         })?;
 
         Ok(())
+    }
+}
+
+#[cfg(test)]
+speculate::speculate! {
+    fn check_args(args: &[&str], expected: Cli) {
+        assert_eq!(
+            expected,
+            Cli::from_clap(&Cli::clap().get_matches_from(args))
+        );
+    }
+
+    fn check_args_err(args: &[&str], error: structopt::clap::ErrorKind) {
+        let result = Cli::clap().get_matches_from_safe(args);
+        assert!(result.is_err());
+        assert_eq!(result.unwrap_err().kind, error);
+    }
+
+    fn s(text: &str) -> String {
+        return text.to_string();
+    }
+
+    describe "CLI" {
+        describe "run subcommand" {
+            it "works with minimal arguments" {
+                check_args(
+                    &["shawl", "run", "--", "foo"],
+                    Cli {
+                        sub: Subcommand::Run {
+                            name: s("Shawl"),
+                            common: CommonOpts {
+                                pass: vec![0],
+                                restart: false,
+                                no_restart: false,
+                                restart_if: vec![],
+                                restart_if_not: vec![],
+                                stop_timeout: 3000,
+                                command: vec![s("foo")],
+                            }
+                        }
+                    },
+                );
+            }
+
+            it "requires a command" {
+                check_args_err(
+                    &["shawl", "run"],
+                    structopt::clap::ErrorKind::MissingRequiredArgument,
+                );
+            }
+
+            it "accepts --pass" {
+                check_args(
+                    &["shawl", "run", "--pass", "1,2", "--", "foo"],
+                    Cli {
+                        sub: Subcommand::Run {
+                            name: s("Shawl"),
+                            common: CommonOpts {
+                                pass: vec![1, 2],
+                                restart: false,
+                                no_restart: false,
+                                restart_if: vec![],
+                                restart_if_not: vec![],
+                                stop_timeout: 3000,
+                                command: vec![s("foo")],
+                            }
+                        }
+                    },
+                );
+            }
+
+            it "accepts --restart" {
+                check_args(
+                    &["shawl", "run", "--restart", "--", "foo"],
+                    Cli {
+                        sub: Subcommand::Run {
+                            name: s("Shawl"),
+                            common: CommonOpts {
+                                pass: vec![0],
+                                restart: true,
+                                no_restart: false,
+                                restart_if: vec![],
+                                restart_if_not: vec![],
+                                stop_timeout: 3000,
+                                command: vec![s("foo")],
+                            }
+                        }
+                    },
+                );
+            }
+
+            it "accepts --no-restart" {
+                check_args(
+                    &["shawl", "run", "--no-restart", "--", "foo"],
+                    Cli {
+                        sub: Subcommand::Run {
+                            name: s("Shawl"),
+                            common: CommonOpts {
+                                pass: vec![0],
+                                restart: false,
+                                no_restart: true,
+                                restart_if: vec![],
+                                restart_if_not: vec![],
+                                stop_timeout: 3000,
+                                command: vec![s("foo")],
+                            }
+                        }
+                    },
+                );
+            }
+
+            it "accepts --restart-if" {
+                check_args(
+                    &["shawl", "run", "--restart-if", "1,2", "--", "foo"],
+                    Cli {
+                        sub: Subcommand::Run {
+                            name: s("Shawl"),
+                            common: CommonOpts {
+                                pass: vec![0],
+                                restart: false,
+                                no_restart: false,
+                                restart_if: vec![1, 2],
+                                restart_if_not: vec![],
+                                stop_timeout: 3000,
+                                command: vec![s("foo")],
+                            }
+                        }
+                    },
+                );
+            }
+
+            it "accepts --restart-if-not" {
+                check_args(
+                    &["shawl", "run", "--restart-if-not", "1,2", "--", "foo"],
+                    Cli {
+                        sub: Subcommand::Run {
+                            name: s("Shawl"),
+                            common: CommonOpts {
+                                pass: vec![0],
+                                restart: false,
+                                no_restart: false,
+                                restart_if: vec![],
+                                restart_if_not: vec![1, 2],
+                                stop_timeout: 3000,
+                                command: vec![s("foo")],
+                            }
+                        }
+                    },
+                );
+            }
+
+            it "accepts --stop-timeout" {
+                check_args(
+                    &["shawl", "run", "--stop-timeout", "500", "--", "foo"],
+                    Cli {
+                        sub: Subcommand::Run {
+                            name: s("Shawl"),
+                            common: CommonOpts {
+                                pass: vec![0],
+                                restart: false,
+                                no_restart: false,
+                                restart_if: vec![],
+                                restart_if_not: vec![],
+                                stop_timeout: 500,
+                                command: vec![s("foo")],
+                            }
+                        }
+                    },
+                );
+            }
+
+            it "accepts --name" {
+                check_args(
+                    &["shawl", "run", "--name", "custom-name", "--", "foo"],
+                    Cli {
+                        sub: Subcommand::Run {
+                            name: s("custom-name"),
+                            common: CommonOpts {
+                                pass: vec![0],
+                                restart: false,
+                                no_restart: false,
+                                restart_if: vec![],
+                                restart_if_not: vec![],
+                                stop_timeout: 3000,
+                                command: vec![s("foo")],
+                            }
+                        }
+                    },
+                );
+            }
+        }
+
+        describe "add subcommand" {
+            it "works with minimal arguments" {
+                check_args(
+                    &["shawl", "add", "--name", "custom-name", "--", "foo"],
+                    Cli {
+                        sub: Subcommand::Add {
+                            name: s("custom-name"),
+                            common: CommonOpts {
+                                pass: vec![0],
+                                restart: false,
+                                no_restart: false,
+                                restart_if: vec![],
+                                restart_if_not: vec![],
+                                stop_timeout: 3000,
+                                command: vec![s("foo")],
+                            }
+                        }
+                    },
+                );
+            }
+
+            it "requires a command" {
+                check_args_err(
+                    &["shawl", "add", "--name", "foo"],
+                    structopt::clap::ErrorKind::MissingRequiredArgument,
+                );
+            }
+
+            it "requires a name" {
+                check_args_err(
+                    &["shawl", "add", "--", "foo"],
+                    structopt::clap::ErrorKind::MissingRequiredArgument,
+                );
+            }
+
+            it "accepts --pass" {
+                check_args(
+                    &["shawl", "add", "--pass", "1,2", "--name", "foo", "--", "foo"],
+                    Cli {
+                        sub: Subcommand::Add {
+                            name: s("foo"),
+                            common: CommonOpts {
+                                pass: vec![1, 2],
+                                restart: false,
+                                no_restart: false,
+                                restart_if: vec![],
+                                restart_if_not: vec![],
+                                stop_timeout: 3000,
+                                command: vec![s("foo")],
+                            }
+                        }
+                    },
+                );
+            }
+
+            it "accepts --restart" {
+                check_args(
+                    &["shawl", "add", "--restart", "--name", "foo", "--", "foo"],
+                    Cli {
+                        sub: Subcommand::Add {
+                            name: s("foo"),
+                            common: CommonOpts {
+                                pass: vec![0],
+                                restart: true,
+                                no_restart: false,
+                                restart_if: vec![],
+                                restart_if_not: vec![],
+                                stop_timeout: 3000,
+                                command: vec![s("foo")],
+                            }
+                        }
+                    },
+                );
+            }
+
+            it "accepts --no-restart" {
+                check_args(
+                    &["shawl", "add", "--no-restart", "--name", "foo", "--", "foo"],
+                    Cli {
+                        sub: Subcommand::Add {
+                            name: s("foo"),
+                            common: CommonOpts {
+                                pass: vec![0],
+                                restart: false,
+                                no_restart: true,
+                                restart_if: vec![],
+                                restart_if_not: vec![],
+                                stop_timeout: 3000,
+                                command: vec![s("foo")],
+                            }
+                        }
+                    },
+                );
+            }
+
+            it "accepts --restart-if" {
+                check_args(
+                    &["shawl", "add", "--restart-if", "1,2", "--name", "foo", "--", "foo"],
+                    Cli {
+                        sub: Subcommand::Add {
+                            name: s("foo"),
+                            common: CommonOpts {
+                                pass: vec![0],
+                                restart: false,
+                                no_restart: false,
+                                restart_if: vec![1, 2],
+                                restart_if_not: vec![],
+                                stop_timeout: 3000,
+                                command: vec![s("foo")],
+                            }
+                        }
+                    },
+                );
+            }
+
+            it "accepts --restart-if-not" {
+                check_args(
+                    &["shawl", "add", "--restart-if-not", "1,2", "--name", "foo", "--", "foo"],
+                    Cli {
+                        sub: Subcommand::Add {
+                            name: s("foo"),
+                            common: CommonOpts {
+                                pass: vec![0],
+                                restart: false,
+                                no_restart: false,
+                                restart_if: vec![],
+                                restart_if_not: vec![1, 2],
+                                stop_timeout: 3000,
+                                command: vec![s("foo")],
+                            }
+                        }
+                    },
+                );
+            }
+
+            it "accepts --stop-timeout" {
+                check_args(
+                    &["shawl", "add", "--stop-timeout", "500", "--name", "foo", "--", "foo"],
+                    Cli {
+                        sub: Subcommand::Add {
+                            name: s("foo"),
+                            common: CommonOpts {
+                                pass: vec![0],
+                                restart: false,
+                                no_restart: false,
+                                restart_if: vec![],
+                                restart_if_not: vec![],
+                                stop_timeout: 500,
+                                command: vec![s("foo")],
+                            }
+                        }
+                    },
+                );
+            }
+        }
+    }
+
+    describe "should_restart_exited_command" {
+        it "handles --restart" {
+            assert!(should_restart_exited_command(&5, true, false, &vec![], &vec![]));
+        }
+
+        it "handles --no-restart" {
+            assert!(!should_restart_exited_command(&0, false, true, &vec![], &vec![]));
+        }
+
+        it "handles --restart-if" {
+            assert!(should_restart_exited_command(&0, false, false, &vec![0], &vec![]));
+            assert!(!should_restart_exited_command(&1, false, false, &vec![0], &vec![]));
+        }
+
+        it "handles --restart-if-not" {
+            assert!(!should_restart_exited_command(&0, false, false, &vec![], &vec![0]));
+            assert!(should_restart_exited_command(&1, false, false, &vec![], &vec![0]));
+        }
+
+        it "restarts nonzero by default" {
+            assert!(!should_restart_exited_command(&0, false, false, &vec![], &vec![]));
+            assert!(should_restart_exited_command(&1, false, false, &vec![], &vec![]));
+        }
+    }
+
+    describe "should_restart_terminated_command" {
+        it "only restarts with --restart" {
+            assert!(!should_restart_terminated_command(false, false));
+            assert!(should_restart_terminated_command(true, false));
+            assert!(!should_restart_terminated_command(false, true));
+        }
     }
 }